Made CacheMemory methods virtual
I am experimenting with modifying caches to have a special tag for
"the cache-line is zero, thus it is not stored in the cache" tag, but
would also like to use normal CacheMemory in some
instances (e.g. instruction cache), so have made the CacheMemory
methods virtual, so that I can extend them in the OO way, while
reusing code.

Change-Id: Ia240964d2b7a67ad25c7b314082b11cc9691aa27