mem-cache: Added getters and setters to CacheBlk
This is so that I can easily track the writes to CacheBlk::data (now
CacheBlk::_data to be in line with the naming conventions).  I had to
create an interface called DataContainer, so that we don't have to
pass `uint8_t*` around (e.g. Packet::trySatisfyFunctional), and that
we can read/write to it.

Change-Id: Idb9d72476a85df95800db35b535395096b31b8f1